apps/backend/requirements.txt
Flask
Flask-Cors
pwdlib[argon2]
pykakasi
google-cloud-translate
janome
python-dotenv
requests
beautifulsoup4
openai
pydantic